fix: run background processes outside/Angular
Provide an `intervalRunner` to ag-grid which will run setTimeout
or interval functions outsize of the Angular zone.

This enables e2e tests to consider pages on which ag-grid-ng2 is present
to settle. Prior to this fix, a page with ag-grid-ng2 could never reach
a state where Protractor would consider it stable for evaluation.

Requires https://github.com/laurelnaiad/ag-grid/commit/889187544d9b9e09c88ddd85a0421d5feccefe61
or similar to be supported in ag-grid.

Closes #43